+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Pending 
+
+* End screen improvements for the full network [#7890](https://github.com/Applifier/unity-ads-webview/pull/7890)
+* WebPlayer MRAID with Disabled AB group [#7938](https://github.com/Applifier/unity-ads-webview/pull/7938)
+* Refactor Privacy Code [#7934](https://github.com/Applifier/unity-ads-webview/pull/7934)
+
 # Mon Jan 28 02:25:46 UTC 2019
 
 * Splitting MRAIDEventHandler to Playable and Programmatic [#7780](https://github.com/Applifier/unity-ads-webview/pull/7780)